pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "myaione-agent"
version = "0.10.0rc7"
description = "MyAIOne Agent — the self-improving AI agent that creates skills from experience, improves them during use, and runs anywhere"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Samuel Mukoti", email = "samuelm@openmy.ai" }]
license = { text = "MIT" }
keywords = [
  "ai", "agent", "llm", "cli", "tui", "myaione", "hermes",
  "qwen", "openai", "anthropic", "openrouter",
  "skills", "tool-calling", "self-improving", "mcp",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Environment :: Console :: Curses",
  "Intended Audience :: Developers",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development",
  "Topic :: Utilities",
]

dependencies = [
  # Core — pinned to known-good ranges to limit supply chain attack surface
  "openai>=2.21.0,<3",
  "anthropic>=0.39.0,<1",
  "python-dotenv>=1.2.1,<2",
  "fire>=0.7.1,<1",
  "httpx[socks]>=0.28.1,<1",
  "rich>=14.3.3,<15",
  "tenacity>=9.1.4,<10",
  "pyyaml>=6.0.2,<7",
  "requests>=2.33.0,<3",  # CVE-2026-25645
  "jinja2>=3.1.5,<4",
  "pydantic>=2.12.5,<3",
  # Interactive CLI (prompt_toolkit is used directly by cli.py)
  "prompt_toolkit>=3.0.52,<4",
  # Tools
  "exa-py>=2.9.0,<3",
  "firecrawl-py>=4.16.0,<5",
  "parallel-web>=0.4.2,<1",
  "fal-client>=0.13.1,<1",
  # Text-to-speech (Edge TTS is free, no API key needed)
  "edge-tts>=7.2.7,<8",
  # Skills Hub (GitHub App JWT auth — optional, only needed for bot identity)
  "PyJWT[crypto]>=2.12.0,<3",  # CVE-2026-32597
]

[project.urls]
Homepage = "https://myai1.ai"
Documentation = "https://docs.myai1.ai"
Repository = "https://github.com/samuelmukoti/myai-agent"
Issues = "https://github.com/samuelmukoti/myai-agent/issues"
"Upstream (Hermes Agent)" = "https://github.com/NousResearch/hermes-agent"

[project.optional-dependencies]
modal = ["modal>=1.0.0,<2"]
daytona = ["daytona>=0.148.0,<1"]
dev = ["debugpy>=1.8.0,<2", "pytest>=9.0.2,<10", "pytest-asyncio>=1.3.0,<2", "pytest-xdist>=3.0,<4", "mcp>=1.2.0,<2"]
messaging = ["python-telegram-bot[webhooks]>=22.6,<23", "discord.py[voice]>=2.7.1,<3", "aiohttp>=3.13.3,<4", "slack-bolt>=1.18.0,<2", "slack-sdk>=3.27.0,<4", "qrcode>=7.0,<8"]
cron = ["croniter>=6.0.0,<7"]
slack = ["slack-bolt>=1.18.0,<2", "slack-sdk>=3.27.0,<4"]
matrix = ["mautrix[encryption]>=0.20,<1", "Markdown>=3.6,<4", "aiosqlite>=0.20", "asyncpg>=0.29"]
cli = ["simple-term-menu>=1.0,<2"]
tts-premium = ["elevenlabs>=1.0,<2"]
voice = [
  # Local STT pulls in wheel-only transitive deps (ctranslate2, onnxruntime),
  # so keep it out of the base install for source-build packagers like Homebrew.
  "faster-whisper>=1.0.0,<2",
  "sounddevice>=0.4.6,<1",
  "numpy>=1.24.0,<3",
]
pty = [
  "ptyprocess>=0.7.0,<1; sys_platform != 'win32'",
  "pywinpty>=2.0.0,<3; sys_platform == 'win32'",
]
honcho = ["honcho-ai>=2.0.1,<3"]
mcp = ["mcp>=1.2.0,<2"]
homeassistant = ["aiohttp>=3.9.0,<4"]
sms = ["aiohttp>=3.9.0,<4"]
acp = ["agent-client-protocol>=0.9.0,<1.0"]
mistral = ["mistralai>=2.3.0,<3"]
bedrock = ["boto3>=1.35.0,<2"]
termux = [
  # Tested Android / Termux path: keeps the core CLI feature-rich while
  # avoiding extras that currently depend on non-Android wheels (notably
  # faster-whisper -> ctranslate2 via the voice extra).
  "python-telegram-bot[webhooks]>=22.6,<23",
  "myaione-agent[cron]",
  "myaione-agent[cli]",
  "myaione-agent[pty]",
  "myaione-agent[mcp]",
  "myaione-agent[honcho]",
  "myaione-agent[acp]",
]
dingtalk = ["dingtalk-stream>=0.20,<1", "alibabacloud-dingtalk>=2.0.0", "qrcode>=7.0,<8"]
feishu = ["lark-oapi>=1.5.3,<2", "qrcode>=7.0,<8"]
web = ["fastapi>=0.104.0,<1", "uvicorn[standard]>=0.24.0,<1"]
# RL training extra — stripped of direct git:// URL deps so the PyPI upload
# isn't rejected.  `atroposlib` (NousResearch/atropos) and `tinker`
# (thinking-machines-lab/tinker) are not on PyPI and can't be referenced as
# direct URLs from a PyPI-distributed package.  Install them yourself if
# you're doing RL work:
#
#   pip install myaione-agent[rl]
#   pip install git+https://github.com/NousResearch/atropos.git@c20c85256e5a45ad31edf8b7276e9c5ee1995a30
#   pip install git+https://github.com/thinking-machines-lab/tinker.git@30517b667f18a3dfb7ef33fb56cf686d5820ba2b
rl = [
  "fastapi>=0.104.0,<1",
  "uvicorn[standard]>=0.24.0,<1",
  "wandb>=0.15.0,<1",
]
# yc-bench extra removed from published metadata for the same reason.
# Install separately if you need the benchmark suite:
#   pip install git+https://github.com/collinear-ai/yc-bench.git@bfb0c88062450f46341bd9a5298903fc2e952a5c
all = [
  "myaione-agent[modal]",
  "myaione-agent[daytona]",
  "myaione-agent[messaging]",
  # matrix: python-olm (required by matrix-nio[e2e]) is upstream-broken on
  # modern macOS (archived libolm, C++ errors with Clang 21+).  On Linux the
  # [matrix] extra's own marker pulls in the [e2e] variant automatically.
  "myaione-agent[matrix]; sys_platform == 'linux'",
  "myaione-agent[cron]",
  "myaione-agent[cli]",
  "myaione-agent[dev]",
  "myaione-agent[tts-premium]",
  "myaione-agent[slack]",
  "myaione-agent[pty]",
  "myaione-agent[honcho]",
  "myaione-agent[mcp]",
  "myaione-agent[homeassistant]",
  "myaione-agent[sms]",
  "myaione-agent[acp]",
  "myaione-agent[voice]",
  "myaione-agent[dingtalk]",
  "myaione-agent[feishu]",
  "myaione-agent[mistral]",
  "myaione-agent[bedrock]",
  "myaione-agent[web]",
]

[project.scripts]
myai = "myai_cli.main:main"
myai-agent = "run_agent:main"
myai-acp = "acp_adapter.entry:main"

[tool.setuptools]
py-modules = ["run_agent", "model_tools", "toolsets", "batch_runner", "trajectory_compressor", "toolset_distributions", "cli", "myai_constants", "myai_state", "myai_time", "myai_logging", "rl_cli", "utils"]

[tool.setuptools.package-data]
myai_cli = ["web_dist/**/*"]

[tool.setuptools.packages.find]
# Include wildcards for packages with subpackages (required once
# `myai_cli/inference/` was added in Phase 3 — without `myai_cli.*`
# that subpackage is silently dropped from built wheels).
#
# Note on `myai_cli.providers`: that's a single flat module file
# (myai_cli/providers.py — the upstream provider catalog) and would
# collide with a `providers/` directory.  Phase 3 code lives under
# `myai_cli/inference/` for that reason.
include = ["agent", "tools", "tools.*", "myai_cli", "myai_cli.*", "gateway", "gateway.*", "tui_gateway", "tui_gateway.*", "cron", "acp_adapter", "plugins", "plugins.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring external services (API keys, Modal, etc.)",
]
addopts = "-m 'not integration' -n auto"
